+++ packages/react-server-next/src/vite/adapters/vercel/entry.ts
@@ -1,3 +1,3 @@
-import { handler } from "@hiogawa/react-server/entry-ssr";
+import { handler } from "@hiogawa/react-server/entry/ssr";
 
 export default handler;
