+++ packages/react-server-next/src/vite/entry-ssr.tsx
@@ -1,4 +1,4 @@
-import { handler } from "@hiogawa/react-server/entry-ssr";
+import { handler } from "@hiogawa/react-server/entry/ssr";
 import { webToNodeHandler } from "@hiogawa/utils-node";
 
 export default webToNodeHandler(handler);
