+++ packages/react-server/examples/basic/src/adapters/vercel-edge.ts
@@ -1,3 +1,3 @@
-import { handler } from "@hiogawa/react-server/entry-ssr";
+import { handler } from "@hiogawa/react-server/entry/ssr";
 
 export default handler;
