+++ packages/react-server/examples/starter/src/adapters/cf.ts
@@ -1,4 +1,4 @@
-import { handler } from "@hiogawa/react-server/entry-ssr";
+import { handler } from "@hiogawa/react-server/entry/ssr";
 
 export default {
   fetch: handler,
