+++ packages/react-server/examples/starter/src/adapters/node.ts
@@ -1,4 +1,4 @@
-import { handler } from "@hiogawa/react-server/entry-ssr";
+import { handler } from "@hiogawa/react-server/entry/ssr";
 import { webToNodeHandler } from "@hiogawa/utils-node";
 
 export default webToNodeHandler(handler);
