+++ packages/react-server/package.json
@@ -10,10 +10,6 @@
     "directory": "packages/react-server"
   },
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
     "./client": {
       "types": "./dist/client.d.ts",
       "default": "./dist/client.js"
@@ -22,37 +18,17 @@
       "types": "./dist/server.d.ts",
       "default": "./dist/server.js"
     },
-    "./runtime-client": {
-      "types": "./dist/runtime-client.d.ts",
-      "default": "./dist/runtime-client.js"
-    },
-    "./runtime-browser": {
-      "types": "./dist/runtime-browser.d.ts",
-      "default": "./dist/runtime-browser.js"
-    },
-    "./runtime-ssr": {
-      "types": "./dist/runtime-ssr.d.ts",
-      "default": "./dist/runtime-ssr.js"
-    },
-    "./runtime-server": {
-      "types": "./dist/runtime-server.d.ts",
-      "default": "./dist/runtime-server.js"
-    },
-    "./entry-browser": {
-      "types": "./dist/entry-browser.d.ts",
-      "default": "./dist/entry-browser.js"
-    },
-    "./entry-ssr": {
-      "types": "./dist/entry-ssr.d.ts",
-      "default": "./dist/entry-ssr.js"
-    },
-    "./entry-server": {
-      "types": "./dist/entry-server.d.ts",
-      "default": "./dist/entry-server.js"
-    },
     "./plugin": {
       "types": "./dist/plugin/index.d.ts",
       "default": "./dist/plugin/index.js"
+    },
+    "./runtime/*": {
+      "types": "./dist/runtime/*.d.ts",
+      "default": "./dist/runtime/*.js"
+    },
+    "./entry/*": {
+      "types": "./dist/entry/*.d.ts",
+      "default": "./dist/entry/*.js"
     }
   },
   "files": ["dist"],
