+++ packages/react-server/src/entry/browser.tsx
@@ -1,8 +1,8 @@
 import { createDebug, memoize, tinyassert } from "@hiogawa/utils";
 import React from "react";
 import ReactDOMClient from "react-dom/client";
-import { initializeReactClientBrowser } from "./features/client-component/browser";
-import { RootErrorBoundary } from "./features/error/error-boundary";
+import { initializeReactClientBrowser } from "../features/client-component/browser";
+import { RootErrorBoundary } from "../features/error/error-boundary";
 import {
   LayoutRoot,
   LayoutStateContext,
@@ -10,22 +10,22 @@ import {
   RouteAssetLinks,
   RouteManifestContext,
   routerRevalidate,
-} from "./features/router/client";
+} from "../features/router/client";
 import {
   Router,
   RouterContext,
   createEncodedBrowserHistory,
   useRouter,
-} from "./features/router/client/router";
+} from "../features/router/client/router";
 import {
   type RouteManifest,
   emptyRouteManifest,
-} from "./features/router/manifest";
-import type { ServerRouterData } from "./features/router/utils";
-import { createStreamRequest } from "./features/server-component/utils";
-import { $__global } from "./global";
-import type { CallServerCallback } from "./types/react";
-import { getFlightStreamBrowser } from "./utils/stream-script";
+} from "../features/router/manifest";
+import type { ServerRouterData } from "../features/router/utils";
+import { createStreamRequest } from "../features/server-component/utils";
+import { $__global } from "../global";
+import type { CallServerCallback } from "../types/react";
+import { getFlightStreamBrowser } from "../utils/stream-script";
 
 const debug = createDebug("react-server:browser");
