+++ packages/react-server/src/features/router/server.tsx
@@ -51,7 +51,7 @@ export function generateRouteModuleTree(globEntries: Record<string, any>) {
 }
 
 // use own "use client" components as external
-function importRuntimeClient(): Promise<typeof import("../../runtime-client")> {
+function importRuntimeClient(): Promise<typeof import("../../runtime/client")> {
   return import("@hiogawa/react-server/runtime-client" as string);
 }
