--- packages/react-server/src/index.ts
@@ -1,1 +1,0 @@
-export {};
