--- packages/react-server/src/runtime-browser.ts
@@ -1,1 +1,0 @@
-export { createServerReference } from "./features/server-action/browser";
