--- packages/react-server/src/runtime-ssr.ts
@@ -1,1 +1,0 @@
-export { createServerReference } from "./features/server-action/ssr";
