+++ packages/react-server/src/runtime/browser.ts
@@ -1,0 +1,1 @@
+export { createServerReference } from "../features/server-action/browser";
