+++ packages/react-server/src/runtime/client.ts
@@ -4,9 +4,9 @@ export {
   ErrorBoundary,
   RedirectBoundary,
   NotFoundBoundary,
-} from "./features/error/error-boundary";
+} from "../features/error/error-boundary";
 export {
   LayoutContent,
   LayoutMatchProvider,
   RemountRoute,
-} from "./features/router/client";
+} from "../features/router/client";
