+++ packages/react-server/src/runtime/server.ts
@@ -1,0 +1,2 @@
+export { registerClientReference } from "../features/client-component/server";
+export { registerServerReference } from "../features/server-action/server";
