+++ packages/react-server/src/runtime/ssr.ts
@@ -1,0 +1,1 @@
+export { createServerReference } from "../features/server-action/ssr";
