+++ packages/react-server/tsup.config.ts
@@ -6,13 +6,13 @@ export default defineConfig([
       "src/index.ts",
       "src/server.ts",
       "src/client.tsx",
-      "src/runtime-client.ts",
-      "src/runtime-browser.ts",
-      "src/runtime-ssr.ts",
-      "src/runtime-server.ts",
-      "src/entry-server.tsx",
-      "src/entry-ssr.tsx",
-      "src/entry-browser.tsx",
+      "src/runtime/server.ts",
+      "src/runtime/client.ts",
+      "src/runtime/browser.ts",
+      "src/runtime/ssr.ts",
+      "src/entry/server.tsx",
+      "src/entry/ssr.tsx",
+      "src/entry/browser.tsx",
       "src/plugin/index.ts",
     ],
     format: ["esm"],
